Wed, May 15, 2024, 3:50 PM -03:00

+++ week2/problems/readability.c
@@ -22,7 +22,7 @@ int main(void)
     printf("%i\n", words);
     printf("%i\n", letters);
 
-    int score1 = (0.0588 * ((letters / words) * 100));
+    int score1 = 0.0588 * ((float) letters / (float) words * 100);
     int score2 = (0.296 * ((sentences / words) * 100));
     printf("%i\n", score1);
     printf("%i\n", score2);
